Move exchange pairs (command & response) into modules
The Command and Response modules were getting out of hand and there was
a great deal of opcode duplication. I found myself implementing the
command and forgetting to make the response function. This brings them
together in harmony while making it harder to see the full set of
opcodes. I’m fine with it here because referencing the opcodes is less
likely in the commands vs decoding opcodes within history or cgm pages.