Fix test expectations for set_base_frequency
It looks like maybe subg_rfspy is returning a 0x01 for each set registry
call. I didn’t see the python code looking for that, but maybe that’s
the cause of the strange syncing issues where sometimes there’s stuff in
the read buffer already when it’s not expected by the code.