Standardize Packet.from_binary/1 return
The convention for Either returns in Erlang is `{:ok, value} | {:error,
reason}`, but `Pummpcomm.Session.Packet.from_binary/1` returns uses
`:invalid_packet` instead of `:error` for the tag.  This both goes
against convention, but also leads to a dialyzer warnings because
`Pummpcomm.Session.PumpExecutor.do_prelude/1`'s `with`'s `else` clause
only handles `{:error, reason}` tuples.  To fix the convention mismatch
the dialyzer warning, change the error format to `{:error,
{:invalid_packet, invalid_packet_reason}}`.  Use a tagged tuple for the
reason is allowed when there's a subheirarchy of errors as is the case
for `:invalid_packet`.

Another convention is to use atoms for the `reason` and not Strings, as
it allows to easier matching.  String errors should only be used when
the error is going directly to the user, so it would make sense for
String errors in the Phoenix or command line UI, but not in the library
called from Phoenix or the command line interface.